MessageID callback
Jumped through a bunch of hoops to implement a callback system to respond back to the command handler with a MessageID once the interaction is posted.

This is needed for voting, so I can store the MessageID with the Vote struct, in order to delete the vote if the message is deleted.

Not the most elegant of hacks, but it should work just fine.