More consistent KVS naming
It was called "sniper" all over the place, but it might not stay sniper for ever.
If I do end up swapping out Sniper for a different KVS, it will annoy me to no end if it's referred to wrong all over the code base, so I started just calling it "kvs" in new code.

Inconsistency annoys me already, so I decided to change it.

Not inside sniper.go itself, obviously, and the data directory is still named the same.