Major refactor of commands/interactions
Everything was sort of spiraling out of control, so I decided to turn commands/ into interactions/ with sub packages for the different types of interaction.
The interaction files remain, though, as I want to group that by function, and not by their individual interactions in that function. For example, the `/faq` stuff and the modal for `/faq add` are all in `interactions/faq.go`, they just use the sub packages as needed.